Move classes to separate source files.
- improving code clarity
