+++ src/CMakeLists.txt
@@ -258,7 +258,6 @@ add_library(fem_lib
     fem/fe_values_views.cc
     fem/mapping_p1.cc
     fem/patch_fe_values.cc
-    fem/patch_point_values.cc
 )
 target_link_libraries(fem_lib
     mesh_lib
