+++ src/fem/patch_point_values.hh
@@ -24,16 +24,14 @@
 #include <Eigen/Dense>
 
 #include "fem/eigen_tools.hh"
-#include "fem/dh_cell_accessor.hh"
-#include "fem/element_values.hh"
+#include "fem/op_function.hh"
 #include "quadrature/quadrature_lib.hh"
 #include "fem/arena_resource.hh"
 #include "fem/arena_vec.hh"
 
 
 template<unsigned int spacedim> class PatchFEValues;
-template<unsigned int spacedim> class PatchOp;
-template <class ValueType> class ElQ;
+template <class ValueType> class ElQ;      // not necessary after merge with patch_fe_values.hh
 template <class ValueType> class FeQ;
 template<unsigned int dim> class BulkValues;
 template<unsigned int dim> class SideValues;
@@ -42,10 +40,6 @@ using Vector = arma::vec3;
 using Tensor = arma::mat33;
 
 
-
-
-/// Type for conciseness
-using ReinitFunction = std::function<void(PatchOp<3> *, IntTableArena &)>;
 
 
 namespace FeBulk {
@@ -122,15 +116,6 @@ namespace FeSide {
         opNItems               ///< Holds number of valid FE operations and value of invalid FE operation
     };
 }
-
-
-/// Distinguishes operations by type and size of output rows
-enum OpSizeType
-{
-	elemOp,      ///< operation is evaluated on elements or sides
-	pointOp,     ///< operation is evaluated on quadrature points
-	fixedSizeOp  ///< operation has fixed size and it is filled during initialization
-};
 
 
 
@@ -545,7 +530,7 @@ public:
      * @param points Allows switched off output of point table,
      * @param ints   Allows switched off output of int (connectivity to elements) table,
      */
-    void print_data_tables(ostream& stream, bool points, bool ints) const {
+    void print_data_tables(std::ostream& stream, bool points, bool ints) const {
         if (points) {
             stream << "Point vals: " << std::endl;
             for (auto *op : operations_) {
@@ -584,7 +569,7 @@ public:
      * Development method.
      * @param bulk_side Needs set 0 (bulk) or 1 (side) for correct output of operation names.
      */
-    void print_operations(ostream& stream, uint bulk_side) const {
+    void print_operations(std::ostream& stream, uint bulk_side) const {
         std::vector< std::vector<std::string> > op_names =
         {
             { "weights", "ref_scalar", "ref_vector", "ref_scalar_grad", "ref_vector_grad", "el_coords", "jacobian", "inv_jac", "jac_det",
@@ -593,12 +578,12 @@ public:
               "side_jac", "side_jac_det", "pt_coords", "JxW", "normal_vec", "scalar_shape", "vector_shape", "grad_scalar_shape", "grad_vector_shape",
               "vector_sym_grad", "vector_divergence" }
         };
-        stream << std::setfill(' ') << " Operation" << setw(12) << "" << "Shape" << setw(2) << ""
-                << "n DOFs" << setw(2) << "" << "Input operations" << endl;
+        stream << std::setfill(' ') << " Operation" << std::setw(12) << "" << "Shape" << std::setw(2) << ""
+                << "n DOFs" << std::setw(2) << "" << "Input operations" << std::endl;
         for (uint i=0; i<operations_.size(); ++i) {
             if (operations_[i] == nullptr) continue;
-            stream << " " << std::left << setw(20) << op_names[bulk_side][i] << "" << " " << setw(6) << operations_[i]->format_shape() << "" << " "
-                << setw(7) << operations_[i]->n_dofs() << "" << " ";
+            stream << " " << std::left << std::setw(20) << op_names[bulk_side][i] << "" << " " << std::setw(6) << operations_[i]->format_shape() << "" << " "
+                << std::setw(7) << operations_[i]->n_dofs() << "" << " ";
             //auto &input_ops = operations_[i]->input_ops();
             for (auto i_o : op_dependency_[i]) stream << op_names[bulk_side][i_o] << " ";
             stream << std::endl;
@@ -607,25 +592,130 @@ public:
 
 protected:
     /// Specialized constructor of zero values object. Do not use in other cases!
-    PatchPointValues(uint dim, std::vector<PatchOp<spacedim> *> &operations, bool is_bulk, PatchFeData &patch_fe_data)
+    PatchPointValues(uint dim, std::vector<PatchOp<spacedim> *> &ref_ops, bool is_bulk, PatchFeData &patch_fe_data)
     : dim_(dim), is_bulk_(is_bulk), elements_map_(300, 0), points_map_(300, 0), patch_fe_data_(patch_fe_data), needs_zero_values_(false) {
         reset();
 
         if (is_bulk) op_dependency_ = std::vector< std::vector<unsigned int> >(FeBulk::BulkOps::opNItems);
         else op_dependency_ = std::vector< std::vector<unsigned int> >(FeSide::SideOps::opNItems);
-        this->create_zero_operations(operations);
 
+        operations_.resize(ref_ops.size(), nullptr);
+        for (uint i_op = 0; i_op < ref_ops.size(); ++i_op ) {
+            auto *op = ref_ops[i_op];
+            if (op == nullptr) continue;
+
+            auto *new_op = make_fe_op(i_op, {op->shape()[0], op->shape()[1]}, &common_reinit::op_base, op->n_dofs(), op->size_type());
+            new_op->allocate_const_result(patch_fe_data_.zero_vec_);
+        }
     }
-
-    void create_zero_operations(std::vector<PatchOp<spacedim> *> &ref_ops);
 
     /// Specialized initialization part of bulk PatchPointValues
     template<unsigned int dim>
-    void init_bulk(uint quad_order);
+    void init_bulk(uint quad_order) {
+        // set dependency of operations
+        this->op_dependency_ = std::vector< std::vector<unsigned int> >(FeBulk::BulkOps::opNItems);
+        this->op_dependency_[FeBulk::BulkOps::opJac] = {FeBulk::BulkOps::opElCoords};
+        this->op_dependency_[FeBulk::BulkOps::opInvJac] = {FeBulk::BulkOps::opJac};
+        this->op_dependency_[FeBulk::BulkOps::opJacDet] = {FeBulk::BulkOps::opJac};
+        this->op_dependency_[FeBulk::BulkOps::opJxW] = {FeBulk::BulkOps::opWeights, FeBulk::BulkOps::opJacDet};
+        this->op_dependency_[FeBulk::BulkOps::opScalarShape] = {FeBulk::BulkOps::opRefScalar};
+        this->op_dependency_[FeBulk::BulkOps::opVectorShape] = {FeBulk::BulkOps::opRefVector};
+        // VectorContravariant: {FeBulk::BulkOps::opRefVector, FeBulk::BulkOps::opJac}
+        // VectorPiola: {FeBulk::BulkOps::opRefVector, FeBulk::BulkOps::opJac, FeBulk::BulkOps::opJacDet}
+        this->op_dependency_[FeBulk::BulkOps::opGradScalarShape] = {FeBulk::BulkOps::opInvJac, FeBulk::BulkOps::opRefScalarGrad};
+        this->op_dependency_[FeBulk::BulkOps::opGradVectorShape] = {FeBulk::BulkOps::opInvJac, FeBulk::BulkOps::opRefVectorGrad};
+        // VectorContravariant: {FeBulk::BulkOps::opInvJac, FeBulk::BulkOps::opRefVectorGrad, FeBulk::BulkOps::opJac}
+        // VectorPiola: {FeBulk::BulkOps::opInvJac, FeBulk::BulkOps::opRefVectorGrad, FeBulk::BulkOps::opJac, FeBulk::BulkOps::opJacDet}
+        this->op_dependency_[FeBulk::BulkOps::opVectorSymGrad] = {FeBulk::BulkOps::opGradVectorShape};
+        this->op_dependency_[FeBulk::BulkOps::opVectorDivergence] = {FeBulk::BulkOps::opGradVectorShape};
+
+        // initialize vectors
+        this->quad_ = new QGauss(dim, 2*quad_order);
+        this->int_sizes_ = {pointOp, pointOp, pointOp};
+        this->operations_.resize(FeBulk::BulkOps::opNItems, nullptr);
+
+        // Create fixed operation
+    	auto *weights = this->make_fixed_op( FeBulk::BulkOps::opWeights, {1}, &common_reinit::op_base );
+        // create result vector of weights operation in assembly arena
+        const std::vector<double> &point_weights_vec = this->quad_->get_weights();
+        weights->allocate_result(point_weights_vec.size(), this->patch_fe_data_.asm_arena_);
+        auto weights_value = weights->result_matrix();
+        for (uint i=0; i<point_weights_vec.size(); ++i)
+            weights_value(0)(i) = point_weights_vec[i];
+
+        // First step: adds element values operations
+        /*auto &el_coords =*/ this->make_new_op( FeBulk::BulkOps::opElCoords, {spacedim, this->dim_+1}, &common_reinit::op_base, OpSizeType::elemOp );
+
+        /*auto &el_jac =*/ this->make_new_op( FeBulk::BulkOps::opJac, {spacedim, this->dim_}, &bulk_reinit::elop_jac<dim>, OpSizeType::elemOp );
+
+        /*auto &el_inv_jac =*/ this->make_new_op( FeBulk::BulkOps::opInvJac, {this->dim_, spacedim}, &bulk_reinit::elop_inv_jac<dim>, OpSizeType::elemOp );
+
+        /*auto &el_jac_det =*/ this->make_new_op( FeBulk::BulkOps::opJacDet, {1}, &bulk_reinit::elop_jac_det<dim>, OpSizeType::elemOp );
+
+        // Second step: adds point values operations
+        /*auto &pt_coords =*/ this->make_new_op( FeBulk::BulkOps::opCoords, {spacedim}, &bulk_reinit::ptop_coords );
+
+        /*auto &JxW =*/ this->make_new_op( FeBulk::BulkOps::opJxW, {1}, &bulk_reinit::ptop_JxW );
+    }
 
     /// Specialized initialization part of side PatchPointValues
     template<unsigned int dim>
-    void init_side(uint quad_order);
+    void init_side(uint quad_order) {
+        // set dependency of operations
+        this->op_dependency_ = std::vector< std::vector<unsigned int> >(FeSide::SideOps::opNItems);
+        this->op_dependency_[FeSide::SideOps::opElJac] = {FeSide::SideOps::opElCoords};
+        this->op_dependency_[FeSide::SideOps::opElInvJac] = {FeSide::SideOps::opElJac};
+        this->op_dependency_[FeSide::SideOps::opSideJac] = {FeSide::SideOps::opSideCoords};
+        this->op_dependency_[FeSide::SideOps::opSideJacDet] = {FeSide::SideOps::opSideJac};
+        this->op_dependency_[FeSide::SideOps::opJxW] = {FeSide::SideOps::opWeights, FeSide::SideOps::opSideJacDet};
+        this->op_dependency_[FeSide::SideOps::opNormalVec] = {FeSide::SideOps::opElInvJac};
+        this->op_dependency_[FeSide::SideOps::opScalarShape] = {FeSide::SideOps::opRefScalar};
+        this->op_dependency_[FeSide::SideOps::opVectorShape] = {FeSide::SideOps::opRefVector};
+        // VectorContravariant: {FeSide::SideOps::opRefVector, FeSide::SideOps::opSideJac}
+        // VectorPiola: {FeSide::SideOps::opRefVector, FeSide::SideOps::opSideJac, FeSide::SideOps::opSideJacDet}
+        this->op_dependency_[FeSide::SideOps::opGradScalarShape] = {FeSide::SideOps::opElInvJac, FeSide::SideOps::opRefScalarGrad};
+        this->op_dependency_[FeSide::SideOps::opGradVectorShape] = {FeSide::SideOps::opElInvJac, FeSide::SideOps::opRefVectorGrad};
+        // VectorContravariant: {FeSide::SideOps::opElInvJac, FeSide::SideOps::opRefVectorGrad, FeSide::SideOps::opElJac}
+        // VectorPiola: {FeSide::SideOps::opElInvJac, FeSide::SideOps::opRefVectorGrad, FeSide::SideOps::opElJac, FeSide::SideOps::opSideJacDet}
+            // TODO ?? define and use opElJacDet
+        this->op_dependency_[FeSide::SideOps::opVectorSymGrad] = {FeSide::SideOps::opGradVectorShape};
+        this->op_dependency_[FeSide::SideOps::opVectorDivergence] = {FeSide::SideOps::opGradVectorShape};
+
+        // initialize vectors
+        this->quad_ = new QGauss(dim-1, 2*quad_order);
+    	this->int_sizes_ = {pointOp, pointOp, pointOp, elemOp, pointOp};
+        this->operations_.resize(FeSide::SideOps::opNItems, nullptr);
+
+        // Create fixed operation
+        auto *weights = this->make_fixed_op( FeSide::SideOps::opWeights, {1},  &common_reinit::op_base ); //lambda_weights );
+        // create result vector of weights operation in assembly arena
+        const std::vector<double> &point_weights_vec = this->quad_->get_weights();
+        weights->allocate_result(point_weights_vec.size(), this->patch_fe_data_.asm_arena_);
+        auto weights_value = weights->result_matrix();
+        for (uint i=0; i<point_weights_vec.size(); ++i)
+            weights_value(0)(i) = point_weights_vec[i];
+
+        // First step: adds element values operations
+        /*auto &el_coords =*/ this->make_new_op( FeSide::SideOps::opElCoords, {spacedim, this->dim_+1}, &common_reinit::op_base, OpSizeType::elemOp );
+
+        /*auto &el_jac =*/ this->make_new_op( FeSide::SideOps::opElJac, {spacedim, this->dim_}, &side_reinit::elop_el_jac<dim>, OpSizeType::elemOp );
+
+        /*auto &el_inv_jac =*/ this->make_new_op( FeSide::SideOps::opElInvJac, {this->dim_, spacedim}, &side_reinit::elop_el_inv_jac<dim>, OpSizeType::elemOp );
+
+        // Second step: adds side values operations
+        /*auto &sd_coords =*/ this->make_new_op( FeSide::SideOps::opSideCoords, {spacedim, this->dim_}, &common_reinit::op_base, OpSizeType::elemOp );
+
+        /*auto &sd_jac =*/ this->make_new_op( FeSide::SideOps::opSideJac, {spacedim, this->dim_-1}, &side_reinit::elop_sd_jac<dim>, OpSizeType::elemOp );
+
+        /*auto &sd_jac_det =*/ this->make_new_op( FeSide::SideOps::opSideJacDet, {1}, &side_reinit::elop_sd_jac_det<dim>, OpSizeType::elemOp );
+
+        // Third step: adds point values operations
+        /*auto &coords =*/ this->make_new_op( FeSide::SideOps::opCoords, {spacedim}, &side_reinit::ptop_coords );
+
+        /*auto &JxW =*/ this->make_new_op( FeSide::SideOps::opJxW, {1}, &side_reinit::ptop_JxW );
+
+        /*auto &normal_vec =*/ this->make_new_op( FeSide::SideOps::opNormalVec, {spacedim}, &side_reinit::ptop_normal_vec<dim>, OpSizeType::elemOp );
+    }
 
     /**
      * Hold integer values of quadrature points of defined operations.
@@ -679,620 +769,6 @@ protected:
     template<unsigned int dim>
     friend class JoinValues;
 };
-
-/**
- * @brief Class represents element or FE operations.
- */
-template<unsigned int spacedim = 3>
-class PatchOp {
-public:
-    /**
-     * Constructor
-     *
-     * Set all data members.
-     */
-    PatchOp(uint dim, std::initializer_list<uint> shape, ReinitFunction reinit_f, OpSizeType size_type,
-            std::vector<PatchOp<spacedim> *> input_ops = {}, uint n_dofs = 1)
-    : dim_(dim), shape_(set_shape_vec(shape)), size_type_(size_type), input_ops_(input_ops),
-      n_dofs_(n_dofs), reinit_func(reinit_f)
-    {}
-
-    /// Aligns shape_vec to 2 items (equal to matrix number of dimensions)
-    std::vector<uint> set_shape_vec(std::initializer_list<uint> shape) const {
-        std::vector<uint> shape_vec(shape);
-        if (shape_vec.size() == 1) shape_vec.push_back(1);
-        ASSERT_EQ(shape_vec.size(), 2);
-        return shape_vec;
-    }
-
-    /**
-     * Return number of operation components
-     *
-     * Value is computed from shape_ vector
-     */
-    inline uint n_comp() const {
-        return shape_[0] * shape_[1];
-    }
-
-    /// Getter for dimension
-    inline uint dim() const {
-        return dim_;
-    }
-
-    /// Getter for size_type_
-    OpSizeType size_type() const {
-        return size_type_;
-    }
-
-    /// Getter for n_dofs_
-    inline uint n_dofs() const {
-        return n_dofs_;
-    }
-
-    /// Return pointer to operation of i_op index in input operation vector.
-    inline PatchOp<spacedim> *input_ops(uint i_op) const {
-        return input_ops_[i_op];
-    }
-
-    /// Getter for shape_
-    inline const std::vector<uint> &shape() const {
-        return shape_;
-    }
-
-    /**
-     * Format shape to string
-     *
-     * Method is used in output development method.
-     */
-    inline std::string format_shape() const {
-        stringstream ss;
-        ss << shape_[0] << "x" << shape_[1];
-        return ss.str();
-    }
-
-    /// Call reinit function on element table if function is defined
-    inline void reinit_function(FMT_UNUSED std::vector<PatchOp<spacedim> *> &operations, IntTableArena &int_table) {
-        reinit_func(this, int_table);
-    }
-
-    inline void allocate_result(size_t data_size, PatchArena &arena) {
-        result_ = Eigen::Vector<ArenaVec<double>, Eigen::Dynamic>(shape_[0] * shape_[1]);
-        for (uint i=0; i<n_comp(); ++i)
-            result_(i) = ArenaVec<double>(data_size, arena);
-    }
-
-    inline void allocate_const_result(ArenaVec<double> &value_vec) {
-        result_ = Eigen::Vector<ArenaVec<double>, Eigen::Dynamic>(shape_[0] * shape_[1]);
-        for (uint i=0; i<n_comp(); ++i)
-            result_(i) = value_vec;
-    }
-
-    /// Return map referenced result as Eigen::Vector
-    Eigen::Map<Eigen::Matrix<ArenaVec<double>, Eigen::Dynamic, Eigen::Dynamic>> result_matrix() {
-        return Eigen::Map<Eigen::Matrix<ArenaVec<double>, Eigen::Dynamic, Eigen::Dynamic>>(result_.data(), shape_[0], shape_[1]);
-    }
-
-    /// Return map referenced result of DOF values as Eigen::Matrix
-    Eigen::Map<Eigen::Matrix<ArenaVec<double>, Eigen::Dynamic, Eigen::Dynamic>> result_sub_matrix(uint i_dof) {
-        ASSERT_LT(i_dof, n_dofs_);
-        uint n_dof_comps = shape_[0] * shape_[1] / n_dofs_;
-        return Eigen::Map<Eigen::Matrix<ArenaVec<double>, Eigen::Dynamic, Eigen::Dynamic>>(result_.data()+i_dof*n_dof_comps, shape_[0], shape_[1] / n_dofs_);
-    }
-
-    /// Return map referenced result as Eigen::Vector
-    Eigen::Vector<ArenaVec<double>, Eigen::Dynamic> &raw_result() {
-        return result_;
-    }
-
-    /// Same as previous but return const reference
-    const Eigen::Vector<ArenaVec<double>, Eigen::Dynamic> &raw_result() const {
-        return result_;
-    }
-
-
-protected:
-    uint dim_;                                    ///< Dimension
-    std::vector<uint> shape_;                     ///< Shape of stored data (size of vector or number of rows and cols of matrix)
-    Eigen::Vector<ArenaVec<double>, Eigen::Dynamic> result_;    ///< Result matrix of operation
-    OpSizeType size_type_;                         ///< Type of operation by size of vector (element, point or fixed size)
-    std::vector<PatchOp<spacedim> *> input_ops_;  ///< Indices of operations in PatchPointValues::operations_ vector on which PatchOp is depended
-    uint n_dofs_;                                 ///< Number of DOFs of FE operations (or 1 in case of element operations)
-
-    ReinitFunction reinit_func;                   ///< Pointer to patch reinit function of element data table specialized by operation
-};
-
-
-/// Defines common functionality of reinit operations.
-struct common_reinit {
-	// empty base operation
-	static inline void op_base(FMT_UNUSED PatchOp<3> * result_op, FMT_UNUSED IntTableArena &el_table) {
-        // empty
-    }
-
-    template<unsigned int dim>
-    static inline void elop_jac(PatchOp<3> * result_op) {
-        // result matrix(spacedim, dim), input matrix(spacedim, dim+1)
-        auto jac_value = result_op->result_matrix();
-        auto coords_value = result_op->input_ops(0)->result_matrix();
-        for (unsigned int i=0; i<3; i++)
-            for (unsigned int j=0; j<dim; j++)
-                jac_value(i,j) = coords_value(i,j+1) - coords_value(i,0);
-    }
-
-    template<unsigned int dim>
-    static inline void elop_inv_jac(PatchOp<3> * result_op) {
-        // result matrix(spacedim, dim), input matrix(spacedim, dim+1)
-        auto inv_jac_value = result_op->result_matrix();
-        auto jac_value = result_op->input_ops(0)->result_matrix();
-        inv_jac_value = eigen_arena_tools::inverse<3, dim>(jac_value);
-    }
-
-    template<unsigned int dim>
-    static inline void elop_jac_det(PatchOp<3> * result_op) {
-        // result double, input matrix(spacedim, dim)
-        auto jac_det_value = result_op->result_matrix();
-        auto jac_value = result_op->input_ops(0)->result_matrix();
-        jac_det_value(0) = eigen_arena_tools::determinant<3, dim>(jac_value).abs();
-    }
-
-    static inline void ptop_JxW(PatchOp<3> * result_op) {
-        auto weights_value = result_op->input_ops(0)->result_matrix();
-        auto jac_det_value = result_op->input_ops(1)->result_matrix();
-        ArenaOVec<double> weights_ovec( weights_value(0,0) );
-        ArenaOVec<double> jac_det_ovec( jac_det_value(0,0) );
-        ArenaOVec<double> jxw_ovec = jac_det_ovec * weights_ovec;
-        auto jxw_value = result_op->result_matrix();
-        jxw_value(0,0) = jxw_ovec.get_vec();
-    }
-
-    /// Common reinit function of vector symmetric gradient on bulk and side points
-    static inline void ptop_vector_sym_grad(PatchOp<3> * result_op, FMT_UNUSED IntTableArena &el_table) {
-        for (uint i_dof=0; i_dof<result_op->n_dofs(); ++i_dof) {
-            Eigen::Map< Eigen::Matrix<ArenaVec<double>, Eigen::Dynamic, Eigen::Dynamic> > grad_vector_dof = result_op->input_ops(0)->result_sub_matrix(i_dof);
-            Eigen::Map< Eigen::Matrix<ArenaVec<double>, Eigen::Dynamic, Eigen::Dynamic> > sym_grad_dof = result_op->result_sub_matrix(i_dof);
-            sym_grad_dof = 0.5 * (grad_vector_dof.transpose() + grad_vector_dof);
-        }
-    }
-	/// Common reinit function of vector divergence on bulk and side points
-    static inline void ptop_vector_divergence(PatchOp<3> * result_op, FMT_UNUSED IntTableArena &el_table) {
-        auto divergence_value = result_op->result_matrix();
-
-        for (uint i_dof=0; i_dof<result_op->n_dofs(); ++i_dof) {
-            Eigen::Map< Eigen::Matrix<ArenaVec<double>, Eigen::Dynamic, Eigen::Dynamic> > grad_vector_dof = result_op->input_ops(0)->result_sub_matrix(i_dof);
-            divergence_value(i_dof) = grad_vector_dof(0,0) + grad_vector_dof(1,1) + grad_vector_dof(2,2);
-        }
-    }
-};
-
-/// Defines reinit operations on bulk points.
-struct bulk_reinit {
-	// element operations
-    template<unsigned int dim>
-    static inline void elop_jac(PatchOp<3> * result_op, FMT_UNUSED IntTableArena &el_table) {
-        common_reinit::elop_jac<dim>(result_op);
-    }
-    template<unsigned int dim>
-    static inline void elop_inv_jac(PatchOp<3> * result_op, FMT_UNUSED IntTableArena &el_table) {
-        common_reinit::elop_inv_jac<dim>(result_op);
-    }
-    template<unsigned int dim>
-    static inline void elop_jac_det(PatchOp<3> * result_op, FMT_UNUSED IntTableArena &el_table) {
-        common_reinit::elop_jac_det<dim>(result_op);
-    }
-
-    // point operations
-    static inline void ptop_coords(FMT_UNUSED PatchOp<3> * result_op, FMT_UNUSED IntTableArena &el_table) {
-        // Implement
-    }
-    static inline void ptop_JxW(PatchOp<3> * result_op, FMT_UNUSED IntTableArena &el_table) {
-        common_reinit::ptop_JxW(result_op);
-    }
-    static inline void ptop_scalar_shape(PatchOp<3> * result_op, FMT_UNUSED IntTableArena &el_table) {
-        auto ref_vec = result_op->input_ops(0)->result_matrix();
-        auto result_vec = result_op->result_matrix();
-
-        uint n_dofs = result_op->n_dofs();
-        uint n_elem = result_vec(0).data_size() / ref_vec(0).data_size();
-
-        ArenaVec<double> elem_vec(n_elem, result_op->raw_result()(0).arena());
-        for (uint i=0; i<n_elem; ++i) {
-            elem_vec(i) = 1.0;
-        }
-        ArenaOVec<double> elem_ovec(elem_vec);
-
-        Eigen::Vector<ArenaOVec<double>, Eigen::Dynamic> ref_ovec(n_dofs);
-        for (uint i=0; i<n_dofs; ++i) {
-            ref_ovec(i) = ArenaOVec<double>( ref_vec(i) );
-        }
-
-        Eigen::Vector<ArenaOVec<double>, Eigen::Dynamic> result_ovec = elem_ovec * ref_ovec;
-        for (uint i=0; i<n_dofs; ++i) {
-            result_vec(i) = result_ovec(i).get_vec();
-        }
-    }
-    static inline void ptop_vector_shape(PatchOp<3> * result_op, FMT_UNUSED IntTableArena &el_table) {
-        auto ref_shape_vec = result_op->input_ops(0)->result_matrix();
-        auto result_vec = result_op->result_matrix(); // result: shape 3x1
-
-        uint n_dofs = result_op->n_dofs();
-        uint n_elem = result_vec(0).data_size() / ref_shape_vec(0).data_size();
-
-        ArenaVec<double> elem_vec(n_elem, result_op->raw_result()(0).arena());
-        for (uint i=0; i<n_elem; ++i) {
-            elem_vec(i) = 1.0;
-        }
-        ArenaOVec<double> elem_ovec(elem_vec);
-
-        Eigen::Matrix<ArenaOVec<double>, Eigen::Dynamic, Eigen::Dynamic> ref_shape_ovec(3, n_dofs);
-        for (uint c=0; c<3*n_dofs; ++c) {
-            ref_shape_ovec(c) = ArenaOVec(ref_shape_vec(c));
-        }
-
-        Eigen::Matrix<ArenaOVec<double>, Eigen::Dynamic, Eigen::Dynamic> result_ovec = elem_ovec * ref_shape_ovec;
-        for (uint c=0; c<3*n_dofs; ++c)
-            result_vec(c) = result_ovec(c).get_vec();
-    }
-    static inline void ptop_vector_contravariant_shape(FMT_UNUSED PatchOp<3> * result_op, FMT_UNUSED IntTableArena &el_table) {
-//        auto &op = operations[vector_sym_grad_op_idx];
-//        auto grad_vector_value = op.input_ops(0).result_matrix();
-//        auto sym_grad_value = op.result_matrix();
-//        sym_grad_value = 0.5 * (grad_vector_value.transpose() + grad_vector_value);
-    }
-    static inline void ptop_vector_piola_shape(FMT_UNUSED PatchOp<3> * result_op, FMT_UNUSED IntTableArena &el_table) {
-//        auto &op = operations[vector_sym_grad_op_idx];
-//        auto grad_vector_value = op.input_ops(0).result_matrix();
-//        auto sym_grad_value = op.result_matrix();
-//        sym_grad_value = 0.5 * (grad_vector_value.transpose() + grad_vector_value);
-    }
-    template<unsigned int dim>
-    static inline void ptop_scalar_shape_grads(PatchOp<3> * result_op, FMT_UNUSED IntTableArena &el_table) {
-        auto inv_jac_vec = result_op->input_ops(0)->result_matrix();    // dim x spacedim=3
-        auto ref_grads_vec = result_op->input_ops(1)->result_matrix();  // dim x n_dofs
-
-        uint n_dofs = result_op->n_dofs();
-
-        Eigen::Matrix<ArenaOVec<double>, Eigen::Dynamic, Eigen::Dynamic> ref_grads_ovec(dim, n_dofs);
-        for (uint i=0; i<dim*n_dofs; ++i) {
-            ref_grads_ovec(i) = ArenaOVec(ref_grads_vec(i));
-        }
-
-        Eigen::Matrix<ArenaOVec<double>, dim, 3> inv_jac_ovec;
-        for (uint i=0; i<dim*3; ++i) {
-            inv_jac_ovec(i) = ArenaOVec(inv_jac_vec(i));
-        }
-
-        auto result_vec = result_op->result_matrix();
-        Eigen::Matrix<ArenaOVec<double>, Eigen::Dynamic, Eigen::Dynamic> result_ovec = inv_jac_ovec.transpose() * ref_grads_ovec;
-        for (uint i=0; i<3*n_dofs; ++i) {
-            result_vec(i) = result_ovec(i).get_vec();
-        }
-    }
-    template<unsigned int dim>
-    static inline void ptop_vector_shape_grads(PatchOp<3> * result_op, FMT_UNUSED IntTableArena &el_table) {
-        auto inv_jac_vec = result_op->input_ops(0)->result_matrix();   // dim x 3
-        auto ref_grads_vec = result_op->input_ops(1)->result_matrix(); // dim x 3*n_dofs
-        auto result_vec = result_op->result_matrix();                  // 3 x 3*n_dofs
-
-        uint n_dofs = result_op->n_dofs();
-
-        Eigen::Matrix<ArenaOVec<double>, dim, 3> inv_jac_ovec;
-        for (uint i=0; i<dim*3; ++i) {
-            inv_jac_ovec(i) = ArenaOVec(inv_jac_vec(i));
-        }
-
-        Eigen::Matrix<ArenaOVec<double>, dim, 3> ref_grads_ovec;
-        for (uint i_dof=0; i_dof<n_dofs; ++i_dof) {
-            for (uint i=0; i<3*dim; ++i) {
-                ref_grads_ovec(i) = ArenaOVec(ref_grads_vec(i_dof*3*dim + i));
-            }
-
-            Eigen::Matrix<ArenaOVec<double>, 3, 3> result_ovec = inv_jac_ovec.transpose() * ref_grads_ovec;
-            for (uint i=0; i<3; ++i) {
-                for (uint j=0; j<3; ++j) {
-                   result_vec(j,i+3*i_dof) = result_ovec(i,j).get_vec();
-                }
-            }
-        }
-    }
-    template<unsigned int dim>
-    static inline void ptop_vector_contravariant_shape_grads(FMT_UNUSED PatchOp<3> * result_op, FMT_UNUSED IntTableArena &el_table) {}
-    template<unsigned int dim>
-    static inline void ptop_vector_piola_shape_grads(FMT_UNUSED PatchOp<3> * result_op, FMT_UNUSED IntTableArena &el_table) {}
-};
-
-
-
-/// Defines reinit operations on side points.
-struct side_reinit {
-	// element operations
-    template<unsigned int dim>
-    static inline void elop_el_jac(PatchOp<3> * result_op, FMT_UNUSED IntTableArena &el_table) {
-        common_reinit::elop_jac<dim>(result_op);
-    }
-    template<unsigned int dim>
-    static inline void elop_el_inv_jac(PatchOp<3> * result_op, FMT_UNUSED IntTableArena &el_table) {
-        common_reinit::elop_inv_jac<dim>(result_op);
-    }
-    template<unsigned int dim>
-    static inline void elop_sd_jac(PatchOp<3> * result_op, FMT_UNUSED IntTableArena &el_table) {
-        // result matrix(spacedim, dim), input matrix(spacedim, dim+1)
-        common_reinit::elop_jac<dim-1>(result_op);
-    }
-    template<unsigned int dim>
-    static inline void elop_sd_jac_det(PatchOp<3> * result_op, FMT_UNUSED IntTableArena &el_table) {
-        common_reinit::elop_jac_det<dim-1>(result_op);
-    }
-
-    // Point operations
-    static inline void ptop_coords(FMT_UNUSED PatchOp<3> * result_op, FMT_UNUSED IntTableArena &el_table) {
-        // Implement
-    }
-    static inline void ptop_JxW(PatchOp<3> * result_op, FMT_UNUSED IntTableArena &el_table) {
-        common_reinit::ptop_JxW(result_op);
-    }
-    template<unsigned int dim>
-    static inline void ptop_normal_vec(PatchOp<3> * result_op, IntTableArena &el_table) {
-        auto normal_value = result_op->result_matrix();
-        auto inv_jac_value = result_op->input_ops(0)->result_matrix();
-        normal_value = inv_jac_value.transpose() * RefElement<dim>::normal_vector_array( el_table(3) );
-
-        ArenaVec<double> norm_vec( normal_value(0).data_size(), normal_value(0).arena() );
-        Eigen::VectorXd A(3);
-        for (uint i=0; i<normal_value(0).data_size(); ++i) {
-            A(0) = normal_value(0)(i);
-            A(1) = normal_value(1)(i);
-            A(2) = normal_value(2)(i);
-            norm_vec(i) = A.norm();
-        }
-
-        for (uint i=0; i<3; ++i) {
-            normal_value(i) = normal_value(i) / norm_vec;
-        }
-    }
-    static inline void ptop_scalar_shape(PatchOp<3> * result_op, IntTableArena &el_table) {
-        auto ref_vec = result_op->input_ops(0)->result_matrix();
-        auto result_vec = result_op->result_matrix();
-
-        uint n_dofs = result_op->n_dofs();
-        uint n_sides = el_table(3).data_size();        // number of sides on patch
-        uint n_patch_points = el_table(4).data_size(); // number of points on patch
-
-        for (uint i_dof=0; i_dof<n_dofs; ++i_dof) {
-            for (uint i_pt=0; i_pt<n_patch_points; ++i_pt) {
-                result_vec(i_dof)(i_pt) = ref_vec(el_table(4)(i_pt), i_dof)(i_pt / n_sides);
-            }
-        }
-    }
-    static inline void ptop_vector_contravariant_shape(FMT_UNUSED PatchOp<3> * result_op, FMT_UNUSED IntTableArena &el_table) {}
-    static inline void ptop_vector_piola_shape(FMT_UNUSED PatchOp<3> * result_op, FMT_UNUSED IntTableArena &el_table) {}
-    static inline void ptop_vector_shape(PatchOp<3> * result_op, IntTableArena &el_table) {
-        auto ref_shape_vec = result_op->input_ops(0)->result_matrix();  // dim+1 x 3*n_dofs
-        auto result_vec = result_op->result_matrix();                   // 3 x n_dofs
-
-        uint n_dofs = result_op->n_dofs();
-        uint n_sides = el_table(3).data_size();
-        uint n_patch_points = el_table(4).data_size();
-
-        for (uint c=0; c<3*n_dofs; c++)
-        	result_vec(c) = ArenaVec<double>(n_patch_points, result_vec(c).arena());
-
-        for (uint i_dof=0; i_dof<n_dofs; ++i_dof) {
-            for (uint i_pt=0; i_pt<n_patch_points; ++i_pt)
-                for (uint c=0; c<3; c++)
-                    result_vec(c,i_dof)(i_pt) = ref_shape_vec(el_table(4)(i_pt),3*i_dof+c)(i_pt / n_sides);
-        }
-    }
-
-template<unsigned int dim>
-    static inline void ptop_scalar_shape_grads(PatchOp<3> * result_op, IntTableArena &el_table) {
-        auto ref_shape_grads = result_op->input_ops(1)->result_matrix();  // dim+1 x dim*n_dofs
-        auto grad_scalar_shape_value = result_op->result_matrix();  // Result vector: 3 x n_dofs
-
-        uint n_dofs = result_op->n_dofs();
-        uint n_points = ref_shape_grads(0).data_size();
-        uint n_sides = el_table(3).data_size();
-        uint n_patch_points = el_table(4).data_size();
-
-        // Expands inverse jacobian to inv_jac_expd_value
-        auto inv_jac_value = result_op->input_ops(0)->result_matrix();
-        Eigen::Matrix<ArenaVec<double>, dim, 3> inv_jac_expd_value;
-        for (uint i=0; i<dim*3; ++i) {
-        	inv_jac_expd_value(i) = ArenaVec<double>( n_patch_points, inv_jac_value(i).arena() );
-        	for (uint j=0; j<n_patch_points; ++j)
-        	    inv_jac_expd_value(i)(j) = inv_jac_value(i)(j%n_sides);
-        }
-
-        // Fill ref shape gradients by q_point. DOF and side_idx
-        Eigen::Matrix<ArenaVec<double>, Eigen::Dynamic, Eigen::Dynamic> ref_shape_grads_expd(dim, n_dofs);
-        for (uint i=0; i<dim*n_dofs; ++i) {
-            ref_shape_grads_expd(i) = ArenaVec<double>( n_patch_points, inv_jac_value(0).arena() );
-        }
-        for (uint i_dof=0; i_dof<n_dofs; ++i_dof) {
-            for (uint i_pt=0; i_pt<n_points; ++i_pt) {
-                uint i_begin = i_pt * n_sides;
-                for (uint i_sd=0; i_sd<n_sides; ++i_sd) {
-                    for (uint i_c=0; i_c<dim; ++i_c) {
-                        ref_shape_grads_expd(i_c, i_dof)(i_begin + i_sd) = ref_shape_grads(el_table(3)(i_sd), i_dof*dim+i_c)(i_pt);
-                    }
-                }
-            }
-        }
-
-        // computes operation result
-        grad_scalar_shape_value = inv_jac_expd_value.transpose() * ref_shape_grads_expd;
-    }
-    template<unsigned int dim>
-    static inline void ptop_vector_shape_grads(PatchOp<3> * result_op, IntTableArena &el_table) {
-        auto result_vec = result_op->raw_result();                        // 3 x 3*n_dofs
-        auto inv_jac_value = result_op->input_ops(0)->result_matrix();    // dim x 3
-        auto ref_vector_grad = result_op->input_ops(1)->result_matrix();  // n_sides*dim x 3*n_dofs
-
-        uint n_dofs = result_op->n_dofs();
-        uint n_points = ref_vector_grad(0).data_size();
-        uint n_patch_sides = el_table(3).data_size();
-        uint n_patch_points = el_table(4).data_size();
-
-        // Expands inverse jacobian to inv_jac_expd_value
-        Eigen::Matrix<ArenaVec<double>, dim, 3> inv_jac_expd_value;
-        for (uint i=0; i<dim*3; ++i) {
-        	inv_jac_expd_value(i) = ArenaVec<double>( n_patch_points, inv_jac_value(i).arena() );
-        	for (uint j=0; j<n_patch_points; ++j)
-        	    inv_jac_expd_value(i)(j) = inv_jac_value(i)(j%n_patch_sides);
-        }
-
-        // Fill ref shape gradients by q_point. DOF and side_idx
-        Eigen::Matrix<ArenaVec<double>, dim, 3> ref_shape_grads_expd;
-        for (uint i=0; i<3*dim; ++i) {
-            ref_shape_grads_expd(i) = ArenaVec<double>( n_patch_points, inv_jac_value(0).arena() );
-        }
-        for (uint i_dof=0; i_dof<n_dofs; ++i_dof) {
-
-            for (uint i_pt=0; i_pt<n_points; ++i_pt) {
-                uint i_begin = i_pt * n_patch_sides;
-                for (uint i_sd=0; i_sd<n_patch_sides; ++i_sd) {
-                    for (uint i_dim=0; i_dim<dim; ++i_dim) {
-                        for (uint i_c=0; i_c<3; ++i_c) {
-                            ref_shape_grads_expd(i_dim, i_c)(i_begin + i_sd) = ref_vector_grad(el_table(3)(i_sd)*dim+i_dim, 3*i_dof+i_c)(i_pt);
-                        }
-                    }
-                }
-            }
-
-            // computes operation result
-            Eigen::Map< Eigen::Matrix<ArenaVec<double>, Eigen::Dynamic, Eigen::Dynamic> > res_submat = result_op->result_sub_matrix(i_dof);
-            res_submat = (inv_jac_expd_value.transpose() * ref_shape_grads_expd).transpose();
-        }
-    }
-    template<unsigned int dim>
-    static inline void ptop_vector_contravariant_shape_grads(FMT_UNUSED PatchOp<3> * result_op, FMT_UNUSED IntTableArena &el_table) {}
-    template<unsigned int dim>
-    static inline void ptop_vector_piola_shape_grads(FMT_UNUSED PatchOp<3> * result_op, FMT_UNUSED IntTableArena &el_table) {}
-};
-
-
-// template specialization
-template<>
-inline void side_reinit::elop_sd_jac<1>(FMT_UNUSED PatchOp<3> * result_op, FMT_UNUSED IntTableArena &el_table) {
-}
-
-template<>
-inline void side_reinit::elop_sd_jac_det<1>(PatchOp<3> * result_op, FMT_UNUSED IntTableArena &el_table) {
-    auto result_vec = result_op->result_matrix();
-    for (uint i=0;i<result_vec(0).data_size(); ++i) {
-        result_vec(0,0)(i) = 1.0;
-    }
-}
-
-
-template<unsigned int spacedim>
-template<unsigned int dim>
-void PatchPointValues<spacedim>::init_bulk(uint quad_order) {
-    // set dependency of operations
-    this->op_dependency_ = std::vector< std::vector<unsigned int> >(FeBulk::BulkOps::opNItems);
-    this->op_dependency_[FeBulk::BulkOps::opJac] = {FeBulk::BulkOps::opElCoords};
-    this->op_dependency_[FeBulk::BulkOps::opInvJac] = {FeBulk::BulkOps::opJac};
-    this->op_dependency_[FeBulk::BulkOps::opJacDet] = {FeBulk::BulkOps::opJac};
-    this->op_dependency_[FeBulk::BulkOps::opJxW] = {FeBulk::BulkOps::opWeights, FeBulk::BulkOps::opJacDet};
-    this->op_dependency_[FeBulk::BulkOps::opScalarShape] = {FeBulk::BulkOps::opRefScalar};
-    this->op_dependency_[FeBulk::BulkOps::opVectorShape] = {FeBulk::BulkOps::opRefVector};
-    // VectorContravariant: {FeBulk::BulkOps::opRefVector, FeBulk::BulkOps::opJac}
-    // VectorPiola: {FeBulk::BulkOps::opRefVector, FeBulk::BulkOps::opJac, FeBulk::BulkOps::opJacDet}
-    this->op_dependency_[FeBulk::BulkOps::opGradScalarShape] = {FeBulk::BulkOps::opInvJac, FeBulk::BulkOps::opRefScalarGrad};
-    this->op_dependency_[FeBulk::BulkOps::opGradVectorShape] = {FeBulk::BulkOps::opInvJac, FeBulk::BulkOps::opRefVectorGrad};
-    // VectorContravariant: {FeBulk::BulkOps::opInvJac, FeBulk::BulkOps::opRefVectorGrad, FeBulk::BulkOps::opJac}
-    // VectorPiola: {FeBulk::BulkOps::opInvJac, FeBulk::BulkOps::opRefVectorGrad, FeBulk::BulkOps::opJac, FeBulk::BulkOps::opJacDet}
-    this->op_dependency_[FeBulk::BulkOps::opVectorSymGrad] = {FeBulk::BulkOps::opGradVectorShape};
-    this->op_dependency_[FeBulk::BulkOps::opVectorDivergence] = {FeBulk::BulkOps::opGradVectorShape};
-
-    // initialize vectors
-    this->quad_ = new QGauss(dim, 2*quad_order);
-    this->int_sizes_ = {pointOp, pointOp, pointOp};
-    this->operations_.resize(FeBulk::BulkOps::opNItems, nullptr);
-
-    // Create fixed operation
-	auto *weights = this->make_fixed_op( FeBulk::BulkOps::opWeights, {1}, &common_reinit::op_base );
-    // create result vector of weights operation in assembly arena
-    const std::vector<double> &point_weights_vec = this->quad_->get_weights();
-    weights->allocate_result(point_weights_vec.size(), this->patch_fe_data_.asm_arena_);
-    auto weights_value = weights->result_matrix();
-    for (uint i=0; i<point_weights_vec.size(); ++i)
-        weights_value(0)(i) = point_weights_vec[i];
-
-    // First step: adds element values operations
-    /*auto &el_coords =*/ this->make_new_op( FeBulk::BulkOps::opElCoords, {spacedim, this->dim_+1}, &common_reinit::op_base, OpSizeType::elemOp );
-
-    /*auto &el_jac =*/ this->make_new_op( FeBulk::BulkOps::opJac, {spacedim, this->dim_}, &bulk_reinit::elop_jac<dim>, OpSizeType::elemOp );
-
-    /*auto &el_inv_jac =*/ this->make_new_op( FeBulk::BulkOps::opInvJac, {this->dim_, spacedim}, &bulk_reinit::elop_inv_jac<dim>, OpSizeType::elemOp );
-
-    /*auto &el_jac_det =*/ this->make_new_op( FeBulk::BulkOps::opJacDet, {1}, &bulk_reinit::elop_jac_det<dim>, OpSizeType::elemOp );
-
-    // Second step: adds point values operations
-    /*auto &pt_coords =*/ this->make_new_op( FeBulk::BulkOps::opCoords, {spacedim}, &bulk_reinit::ptop_coords );
-
-    /*auto &JxW =*/ this->make_new_op( FeBulk::BulkOps::opJxW, {1}, &bulk_reinit::ptop_JxW );
-}
-
-
-template<unsigned int spacedim>
-template<unsigned int dim>
-void PatchPointValues<spacedim>::init_side(uint quad_order) {
-    // set dependency of operations
-    this->op_dependency_ = std::vector< std::vector<unsigned int> >(FeSide::SideOps::opNItems);
-    this->op_dependency_[FeSide::SideOps::opElJac] = {FeSide::SideOps::opElCoords};
-    this->op_dependency_[FeSide::SideOps::opElInvJac] = {FeSide::SideOps::opElJac};
-    this->op_dependency_[FeSide::SideOps::opSideJac] = {FeSide::SideOps::opSideCoords};
-    this->op_dependency_[FeSide::SideOps::opSideJacDet] = {FeSide::SideOps::opSideJac};
-    this->op_dependency_[FeSide::SideOps::opJxW] = {FeSide::SideOps::opWeights, FeSide::SideOps::opSideJacDet};
-    this->op_dependency_[FeSide::SideOps::opNormalVec] = {FeSide::SideOps::opElInvJac};
-    this->op_dependency_[FeSide::SideOps::opScalarShape] = {FeSide::SideOps::opRefScalar};
-    this->op_dependency_[FeSide::SideOps::opVectorShape] = {FeSide::SideOps::opRefVector};
-    // VectorContravariant: {FeSide::SideOps::opRefVector, FeSide::SideOps::opSideJac}
-    // VectorPiola: {FeSide::SideOps::opRefVector, FeSide::SideOps::opSideJac, FeSide::SideOps::opSideJacDet}
-    this->op_dependency_[FeSide::SideOps::opGradScalarShape] = {FeSide::SideOps::opElInvJac, FeSide::SideOps::opRefScalarGrad};
-    this->op_dependency_[FeSide::SideOps::opGradVectorShape] = {FeSide::SideOps::opElInvJac, FeSide::SideOps::opRefVectorGrad};
-    // VectorContravariant: {FeSide::SideOps::opElInvJac, FeSide::SideOps::opRefVectorGrad, FeSide::SideOps::opElJac}
-    // VectorPiola: {FeSide::SideOps::opElInvJac, FeSide::SideOps::opRefVectorGrad, FeSide::SideOps::opElJac, FeSide::SideOps::opSideJacDet}
-        // TODO ?? define and use opElJacDet
-    this->op_dependency_[FeSide::SideOps::opVectorSymGrad] = {FeSide::SideOps::opGradVectorShape};
-    this->op_dependency_[FeSide::SideOps::opVectorDivergence] = {FeSide::SideOps::opGradVectorShape};
-
-    // initialize vectors
-    this->quad_ = new QGauss(dim-1, 2*quad_order);
-	this->int_sizes_ = {pointOp, pointOp, pointOp, elemOp, pointOp};
-    this->operations_.resize(FeSide::SideOps::opNItems, nullptr);
-
-    // Create fixed operation
-    auto *weights = this->make_fixed_op( FeSide::SideOps::opWeights, {1},  &common_reinit::op_base ); //lambda_weights );
-    // create result vector of weights operation in assembly arena
-    const std::vector<double> &point_weights_vec = this->quad_->get_weights();
-    weights->allocate_result(point_weights_vec.size(), this->patch_fe_data_.asm_arena_);
-    auto weights_value = weights->result_matrix();
-    for (uint i=0; i<point_weights_vec.size(); ++i)
-        weights_value(0)(i) = point_weights_vec[i];
-
-    // First step: adds element values operations
-    /*auto &el_coords =*/ this->make_new_op( FeSide::SideOps::opElCoords, {spacedim, this->dim_+1}, &common_reinit::op_base, OpSizeType::elemOp );
-
-    /*auto &el_jac =*/ this->make_new_op( FeSide::SideOps::opElJac, {spacedim, this->dim_}, &side_reinit::elop_el_jac<dim>, OpSizeType::elemOp );
-
-    /*auto &el_inv_jac =*/ this->make_new_op( FeSide::SideOps::opElInvJac, {this->dim_, spacedim}, &side_reinit::elop_el_inv_jac<dim>, OpSizeType::elemOp );
-
-    // Second step: adds side values operations
-    /*auto &sd_coords =*/ this->make_new_op( FeSide::SideOps::opSideCoords, {spacedim, this->dim_}, &common_reinit::op_base, OpSizeType::elemOp );
-
-    /*auto &sd_jac =*/ this->make_new_op( FeSide::SideOps::opSideJac, {spacedim, this->dim_-1}, &side_reinit::elop_sd_jac<dim>, OpSizeType::elemOp );
-
-    /*auto &sd_jac_det =*/ this->make_new_op( FeSide::SideOps::opSideJacDet, {1}, &side_reinit::elop_sd_jac_det<dim>, OpSizeType::elemOp );
-
-    // Third step: adds point values operations
-    /*auto &coords =*/ this->make_new_op( FeSide::SideOps::opCoords, {spacedim}, &side_reinit::ptop_coords );
-
-    /*auto &JxW =*/ this->make_new_op( FeSide::SideOps::opJxW, {1}, &side_reinit::ptop_JxW );
-
-    /*auto &normal_vec =*/ this->make_new_op( FeSide::SideOps::opNormalVec, {spacedim}, &side_reinit::ptop_normal_vec<dim>, OpSizeType::elemOp );
-}
-
 
 
 #endif /* PATCH_POINT_VALUES_HH_ */
